refactor(dx): diagnostic interaction を app module に分離
Refs #310
