+++ src/app/mod.rs
@@ -1,3 +1,4 @@
 //! アプリケーション層 (Slint UI を駆動するロジック)。
 
+pub(crate) mod diagnostics;
 pub(crate) mod diff_viewer;
